refactor: break circular dependency between plugin and composables
Extract ColorScheme type and DefaultColorSchemeKey to src/types/theme.ts
so both plugin/index.ts and composables/use-theme.ts depend on the shared
types module instead of importing from each other.

Closes #2
